requirements.txt
Flask
python-dotenv
openai
PyPDF2
werkzeug
matplotlib  # Added for graph generation
